pyproject.toml
[project]
name = "py_load_eudravigilance"
version = "0.1.0"
description = "High-performance ETL for EudraVigilance ICSRs into relational databases."
authors = [
    {name = "Gowtham Rao", email = "rao@ohdsi.org"},
]
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "lxml>=4.9.3",
    "pyyaml>=6.0.1",
    "fsspec>=2023.10.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pytest-mock (>=3.15.0,<4.0.0)",
    "sqlalchemy>=2.0.22",
]

[project.optional-dependencies]
postgres = ["psycopg2-binary>=2.9.9", "sqlalchemy>=2.0.22"]
redshift = ["redshift_connector", "boto3"]
databricks = ["databricks-sql-connector", "boto3"]
bigquery = ["google-cloud-bigquery", "google-cloud-storage"]
s3 = ["s3fs>=2023.10.0"]
gcs = ["gcsfs>=2023.10.0"]
azure = ["azure-storage-blob>=12.18.2"]
cli = ["typer>=0.9.0"]
all = [
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.22",
    "redshift_connector",
    "boto3",
    "databricks-sql-connector",
    "google-cloud-bigquery",
    "google-cloud-storage",
    "s3fs>=2023.10.0",
    "gcsfs>=2023.10.0",
    "azure-storage-blob>=12.18.2",
    "typer>=0.9.0",
]
dev = [
    "pytest>=7.4.2",
    "ruff>=0.1.0",
    "black>=23.9.1",
    "isort>=5.12.0",
    "mypy>=1.5.1",
    "pytest-cov>=4.1.0",
    "testcontainers[postgres]>=3.7.1",
    "mkdocs>=1.5.3",
    "mkdocs-material>=9.4.6",
]

[project.scripts]
eudravigloader = "py_load_eudravigilance.cli:app"

[project.entry-points."py_load_eudravigilance.loaders"]
postgresql = "py_load_eudravigilance.postgres:PostgresLoader"
redshift = "py_load_eudravigilance.redshift:RedshiftLoader"
databricks = "py_load_eudravigilance.databricks:DatabricksLoader"
bigquery = "py_load_eudravigilance.bigquery:BigQueryLoader"

[tool.poetry]
packages = [{include = "py_load_eudravigilance", from = "src"}]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.main"

[tool.ruff]
line-length = 88
select = ["E", "W", "F", "I", "C", "B"]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
